Malagon_MRCapplication/requirements.txt
mysql-connector-python>=8.0.32
tkcalendar>=1.6.1